Make noStyles work also without preprocessors
Hi. Thanks for a nice project!

I have some problems with deasync (I think), it just hangs forever when using a preprocessor. But I only need preprocessing for styles, and just removing the styles would work.

However because of what I think might be a small bug, `noStyles` when there are no preprocessors. This PR fixes that.